(Test): Unit test for base mock template - no real data [pv622][at2222][tsv22]
