+++ aignostic/tests/modelAPI/model/scikit_mock.py
@@ -9,13 +9,19 @@ app = FastAPI()
 
 model : Pipeline = pickle.load(open('scikit_model.sav', 'rb'))
 
-@app.post("/query_all")
+@app.post("/predict")
 def predict(dataset : DataSet) -> QueryOutput:
     """
     Given a dataset, predict the expected outputs for the model
     """
     # Return empty dataframe for now - fill this in with actual test models when trained
-    return QueryOutput(model.predict(dataset.to_dataframe()))
+    return QueryOutput(columns={})
+    # return QueryOutput(columns=model.predict(to_dataframe(dataset))).to_dict()
+
+@app.get("/hello")
+def hello() -> str:
+    print("Hello World")
+    return "Hello World"
 
 """
 TODO: (Low Priority) Extend to batch querying / single datapoint querying for convenience
